currency_converter/pyproject.toml
[project]
name = "currency-converter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]

[tool.uv.workspace]
members = [
    "currency-converter",
]
